+++ lib/TonePlayer/TonePlayer.h
@@ -8,13 +8,12 @@ class TonePlayer
     public:
         struct Tone
         {
-            unsigned long freq;
-            unsigned long endAt;
-
+            unsigned short int freq;
+            unsigned short int endAt;
         };
         TonePlayer(ConfigStorage* _configStorage, uint8_t _buzzerPin);
-        void loadTones(const Tone* _tones, const unsigned short int _tonesLength, bool _playLoop = false, bool _tonePriority = false);
-        void playTones(const Tone* _tones, const unsigned short int _tonesLength, bool _playLoop = false, bool _tonePriority = false);
+        void loadTones(const Tone* _tones, const uint8_t _tonesLength, bool _playLoop = false, bool _tonePriority = false);
+        void playTones(const Tone* _tones, const uint8_t _tonesLength, bool _playLoop = false, bool _tonePriority = false);
         bool play();
         void stop();
     private:
@@ -23,7 +22,7 @@ class TonePlayer
         bool playLoop; 
         bool tonePriority; 
         unsigned long startAt;
-        unsigned short int tonesLength = 0; // how many tones in a melody
+        uint8_t tonesLength = 0; // how many tones in a melody
         const static unsigned short int nonPriorityToneDelay = 40;
         const Tone* tones;
 };
